+++ features/auth/ui/Register/ModalWindow/ModalWindow.module.scss
@@ -15,10 +15,9 @@
   width: 378px;
   height: 228px;
   box-sizing: border-box;
-  border: 1px solid var(--color-dark-100);
+  border: 1px solid rgb(76 76 76);
   border-radius: 2px;
-  background: var(--color-dark-300);
-  color: var(--color-light-100);
+  background: rgb(51 51 51);
 }
 
 .header {
@@ -48,12 +47,14 @@
 }
 
 .title {
+  color: white;
   margin: 0;
   align-content: center;
 }
 
 .text {
-  font-size: var(--text-md);
-  font-weight: var(--font-regular);
+  color: white;
+  font-size: 16px;
+  font-weight: 400;
   margin-bottom: 18px;
 }
