+++ features/auth/ui/Register/ModalWindow/ModalWindow.tsx
@@ -1,8 +1,8 @@
 // "use client"
-import CloseIcon from "@/shared/assets/svg/Close.svg"
+import CloseIcon from "../../assets/svg/Close.svg"
 import Button from "@/shared/ui/button/button"
 import { Dialog } from "@radix-ui/react-dialog"
-import s from "./ModalWindow.module.css"
+import s from "./ModalWindow.module.scss"
 import Image from "next/image"
 
 type Props = {
