+++ features/auth/ui/Register/RegisterForm/RegisterForm.module.scss
