+++ package.json
@@ -15,8 +15,6 @@
     "test:coverage": "jest --coverage"
   },
   "dependencies": {
-    "@radix-ui/react-checkbox": "^1.3.2",
-    "@radix-ui/react-icons": "^1.3.2",
     "@radix-ui/themes": "^3.2.1",
     "@reduxjs/toolkit": "^2.8.2",
     "next": "15.3.5",
