+++ shared/assets/icons/components/EyeIcon.tsx
