--- shared/ui/checkbox/checkbox.module.css
@@ -1,68 +1,0 @@
-.container {
-  display: flex;
-  align-items: center;
-  margin-left: 20px;
-}
-
-.checkbox {
-  position: relative;
-  width: 24px;
-  height: 24px;
-  border: none;
-  background: none;
-  padding: 0;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  z-index: 1;
-
-  &:hover,
-  &:active {
-    &::before {
-      content: "";
-      width: calc(100% + 12px);
-      height: calc(100% + 12px);
-      position: absolute;
-      top: 50%;
-      left: 50%;
-      z-index: -1;
-      transform: translate(-50%, -50%);
-      background: var(--color-dark-100);
-      border-radius: 50%;
-    }
-  }
-
-  &:focus {
-    &::after {
-      display: none;
-    }
-
-    &::before {
-      content: "";
-      width: calc(100% + 12px);
-      height: calc(100% + 12px);
-      position: absolute;
-      top: 50%;
-      left: 50%;
-      z-index: -1;
-      transform: translate(-50%, -50%);
-      background: var(--color-dark-500);
-      border-radius: 50%;
-    }
-  }
-
-  &:disabled {
-    &::before {
-      display: none;
-    }
-
-    && + .span {
-      color: var(--color-light-900);
-    }
-  }
-}
-
-.span {
-  margin-left: 8px;
-  color: var(--color-light-100);
-}
