--- widgets/auth/ui/Signup/Signup.module.css
@@ -1,4 +1,0 @@
-.page {
-  padding: 24px 0;
-  background-color: var(--color-dark-700);
-}
